+++ Backend/sources/apps/app/app_views.py
@@ -1,18 +1,21 @@
 from django.shortcuts import render
-from rest_framework.viewsets import ModelViewSet
-from rest_framework.generics import RetrieveAPIView, CreateAPIView, ListAPIView
+from rest_framework.views import APIView
+from rest_framework.generics import RetrieveAPIView, CreateAPIView, ListAPIView, UpdateAPIView, GenericAPIView
 from rest_framework.decorators import action
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated, AllowAny
 from .controllers.execute import processCalculo
 from rest_framework.authentication import TokenAuthentication, SessionAuthentication
-
+from django.utils import timezone
+from django.db.models import Sum
+from django.shortcuts import get_object_or_404
 #Models
 from . import models
 from . import app_serializers
+from datetime import datetime
 # Create your views here.
- 
+from . import enums
 ## Calculo Views
 class Calculo(RetrieveAPIView):
     permission_classes = [AllowAny]
@@ -27,6 +30,47 @@ class Calculo(RetrieveAPIView):
         except Exception as exception:
             #Do logger class for this
             pass
+
+## Register views
+
+class RegisterRetrieveAPIView(RetrieveAPIView):
+    queryset = models.Register.objects.all()
+    serializer_class = app_serializers.RegisterRetrieveSerializer
+
+    def get_object(self):
+        
+        queryset = self.get_queryset()
+        obj = get_object_or_404(queryset, date=datetime.now().strftime("%Y-%m-%d"))
+        return obj
+
+
+class RegisterCreateAPIView(CreateAPIView):
+    serializer_class = app_serializers.RegisterCreateSerializer
+
+    def create(self, request, *args, **kwargs):
+
+        date = datetime.now().strftime("%Y-%m-%d")
+        # print(date, flush=True)
+        register = models.Register.objects.filter(date = date).first()
+        
+        if (register):
+            print("existe",flush=True)
+            visited = register.visited + 1
+            data = {"visited":visited}
+            serializer = self.get_serializer(register, data = data, partial=True)
+
+            if(serializer.is_valid(raise_exception=True)):
+
+                serializer.save()
+                return Response({}, status=status.HTTP_201_CREATED)
+
+        data = {"date": date}
+        serializer = self.get_serializer(data = data)
+        serializer.is_valid(raise_exception=True)
+        serializer.save()
+
+        return Response({}, status=status.HTTP_201_CREATED)
+
 
 ## Chat Views
 
@@ -49,6 +93,8 @@ class ChatRetrieveAPIView(RetrieveAPIView):
     queryset = models.Chat.objects.all()
     serializer_class = app_serializers.ChatSerializer
 
+    
+
 
 class ChatListAPIView(ListAPIView):
     # authentication_classes = [SessionAuthentication, TokenAuthentication]
@@ -56,3 +102,25 @@ class ChatListAPIView(ListAPIView):
 
     queryset = models.Chat.objects.all()
     serializer_class = app_serializers.ChatSerializer
+
+class ChatCounterRetrieveAPIView(RetrieveAPIView):
+
+    def get(self, request):
+
+        try:
+            data = {
+                "chats_on": models.Chat.objects.filter(status=enums.ChatStatus.Open).count(),
+                "chats_off": models.Chat.objects.filter(status=enums.ChatStatus.Closed).count()
+            }
+
+            return Response(data, status=status.HTTP_200_OK)
+        
+        except Exception as e:
+            return Response(data, status=status.HTTP_400_BAD_REQUEST)
+
+
+
+
+        
+
+