+++ Backend/sources/apps/app/models.py
@@ -5,6 +5,8 @@ from .enums import ChatStatus
 from channels.layers import get_channel_layer
 from asgiref.sync import async_to_sync
 from .enums import WSNotificationType
+from django.conf import settings
+import pytz 
 # Create your models here.
 #chat 1----* messages 
 #cliente 1 --- 1 chat
@@ -57,3 +59,9 @@ class Client(models.Model):
     #Metadata
     #Methods
 
+
+class Register(models.Model):
+    date = models.DateField()
+    visited = models.PositiveIntegerField(default=1)
+
+