+++ Frontend/src/app/_components/admin/stats/interfaces/stats.interface.ts
@@ -1,0 +1,5 @@
+export interface IStats {
+    date: string
+    visited: number
+    total_visits: number
+}