+++ Frontend/src/app/_components/admin/stats/stats.tsx
@@ -1,14 +1,20 @@
+import Values from "./values";
+
 export default function Stats(){
 
-    return (
-        <section className="w-full h-full bg-[#1a1818] p-20 rounded-xl">
-            <section className="w-full h-full flex gap-2 items-center justify-center">
-                <div className=" w-28 h-28 border-b-2"></div>
-                <div className=" w-28 h-28 border-b-2"></div>
-                
 
-            </section>
-            
+    return (
+        <section className="w-full h-full bg-[#1a1818] rounded-xl flex flex-col ">
+           
+       
+            <div className="w-full h-16 bg-slate-600 shadow-[4.0px_8.0px_8.0px_rgba(0,0,0,0.38)] overflow-hidden flex justify-center items-center">
+                <h2 className="text-center text-white text-2xl font-bold">Panel administrativo</h2>
+            </div>
+   
+        
+            <Values/>
+    
+   
         </section>
     )
 }