+++ Frontend/src/app/_core/services/httpRequest.ts
@@ -1,5 +1,5 @@
 import { createContext } from "react";
-import { IHttpOptions } from "../interfaces/core.interface";
+import { IHTTPdetail, IHTTPresponse, IHttpOptions } from "../interfaces/core.interface";
 
 
 class httpRequestService{
@@ -13,16 +13,16 @@ class httpRequestService{
 
         return options
     }
- 
-    async request(method: string, options: IHttpOptions, with_credentials: boolean = false): Promise<[number, any]>{
+    //Promise<[number, any]>
+    async request<T>(method: string, options: IHttpOptions, with_credentials: boolean = false): Promise< IHTTPresponse<T> >{
 
         let settings: any = {
             method: method,
             headers: {
                 "Content-Type": "application/json"
-            },
-            credentials: 'include'
+            }
         }
+        
         if (with_credentials === true){
             settings.credentials = 'include'
         }
@@ -35,14 +35,17 @@ class httpRequestService{
         try{
             const response = await fetch(options.url, settings)
             const status = response.status
-            const data = await response.json()
-            
+            const data: T = await response.json()
             
 
-            return [status,data]
+
+            // return [status,data]
+            return {status: status,data: data}
     
         } catch (error){
-            return [500, { error: 'Internal Server Error' }];
+            //return [500, { error: 'Internal Server Error' }];
+            return {status: 500, data: undefined}
+            
         }
     }
 }